+++ index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Harambot</title>
+
+    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.1/css/bootstrap.min.css">
+    <style>
+        body { padding-top:50px; }
+    </style>
+</head>
+<body>
+    <div class="container"> 	
+		<form action="/" method="post">
+			<h1>Harambot</h1>
+			<label>Username</label>
+			<input name="username" type="text"></input>
+			<label>Password</label>
+			<input name="password" type="password"></input>
+			<label>Thread ID</label>
+			<input name="threadId"></input>
+			<button name="downloadData" type="submit" value="Download Data">Download Data</button>	
+		</form>
+    </div>
+</body>
+</html>
